docformatter/requirements.txt
PyQt6 >= 6.6.0
python-docx >= 1.1.0
lxml >= 5.0.0
